backend/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
Pillow>=10.0.0
python-decouple>=3.8
dj-database-url>=2.0.0
django-extensions>=3.2.0
django-debug-toolbar>=4.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0